+++ src/main/java/com/qa/rest/PlayController.java
@@ -1,0 +1,4 @@
+package com.qa.rest;
+
+public class PlayController {
+}
