+++ src/main/java/com/qa/rest/PlaybookController.java
@@ -1,0 +1,4 @@
+package com.qa.rest;
+
+public class PlaybookController {
+}
